refactor: unify DockerExecutor type and interface for improved abstraction
